fix: GraphQL points Float type + ensure integer redemption values

+++ graphql-gateway/src/modules/rewards/reward.type.js
@@ -10,7 +10,7 @@ const RewardType = gql`
     id: ID
     userId: Int
     orderId: Int
-    points: Int
+    points: Float
     type: String
     description: String
     createdAt: String
